DATACASS-718 - Fix Integer Overflow in Update.toString().
Original pull request: #169.

+++ spring-data-cassandra/src/main/java/org/springframework/data/cassandra/core/query/Update.java
@@ -37,6 +37,7 @@ import org.springframework.util.StringUtils;
  * </pre>
  *
  * @author Mark Paluch
+ * @author Chema Vinacua 
  * @since 2.0
  */
 public class Update {
@@ -662,7 +663,7 @@ public class Update {
 		@Override
 		public String toString() {
 			return String.format("%s = %s %s %d", getColumnName(), getColumnName(), value.doubleValue() > 0 ? "+" : "-",
-					Math.abs(value.intValue()));
+					Math.abs(value.longValue()));
 		}
 	}
